update delete s3 key

+++ api/restful/s3_api.py
@@ -80,8 +80,9 @@ def delete_wiki(request):
     response_data = {}
     if request.method == 'POST':
         key_name = request.POST['key_name'] or None
-        if key_name:
+        if key_name and s3.is_file_exist(key_name):
             s3.delete_file(key_name)
             return Response(data=response_data, status=status.HTTP_200_OK)
+        response_data['error'] = 'Invalid key name'
         return Response(data=response_data, status=status.HTTP_400_BAD_REQUEST)
     return Response(data=response_data, status=status.HTTP_405_METHOD_NOT_ALLOWED)
